Introduce a new foundation for Dev Services like experience
The basic idea behind this is that the LLM model being run and the
API being exposed is not always tightly couple - for example
the Mistral models can be run in Ollama, not just in Mistral's
SaaS offering.
This PR sets the foundation for having an inference server
(for now only Ollama) be able to run inference for
the LLM models the user configured.
In the Ollama implementation we are able to instruct Ollama
to pull the selected model name(if it exists and if
it is not already present) and make the necessary
configuration behind the scenes so each configured LangChain4j
chat model configuration will use the inference server.

The following remains to be done:
* Use the same idea for the rest of the LangChain4j models
(like embedding models)
* Use Ollama's OpenAI compatibility mode to support the OpenAI extension