Also allow to inject PgVectorEmbeddingStore with upstream type
Our PgVectorEmbeddingStore extends the one from LangChain4j and when
injecting, it's easy to choose the one from upstream and then you're
stuck with injection errors that are a bit hard to figure out as
everything looks fine.

Let's make it a bit easier as it doesn't cost us anything.