Introduce AiService class and method names into GenAI metrics
This is done using the Duplicated Context as currently LangChain4j
doesn't provide a more fine-grained way of providing information
to the ChatModelListener.
Nevertheless, the use of Duplicated Context is done in way that
ensures that the tags will only be added if the Duplicated Context
exists. If they don't, metrics continue to work.