Refactor upstream guardrails
Refactor to use upstream guardrails.

For now I'm leaving the existing Quarkus-specific implementation in place, but fully deprecated. We can decide if we'd like to rip it out as part of this PR, or have both implementations for a little bit of time in order to give people time to transition.

Closes #1284